Refactor the device tests to avoid duplicate tests
* Several duplicated tests because the HandlerTestBase class was re-used for the same control
* There is now a new HandlerTestBasement with _just_ helper methods for all handler tests
* Refactored the set of cross-handler tests for Focus, TextInput and TextStyle to depend on HandlerTestBasement
