+++ src/Controls/tests/DeviceTests/Elements/Border/BorderTests.cs
@@ -8,7 +8,6 @@ using Xunit;
 
 namespace Microsoft.Maui.DeviceTests
 {
-#if !TIZEN
 	[Category(TestCategory.Border)]
 	public partial class BorderTests : ControlsHandlerTestBase
 	{
@@ -33,5 +32,4 @@ namespace Microsoft.Maui.DeviceTests
 			await AssertColorAtPoint(border, expected, typeof(BorderHandler), 10, 10);
 		}
 	}
-#endif
 }
