+++ src/Core/tests/DeviceTests/Handlers/Button/ButtonFocusTests.cs
@@ -1,0 +1,12 @@
+﻿using Microsoft.Maui.DeviceTests.Stubs;
+
+namespace Microsoft.Maui.DeviceTests
+{
+	[Category(TestCategory.Button)]
+	public class ButtonFocusTests : FocusHandlerTests<ButtonHandler, ButtonStub, VerticalStackLayoutStub>
+	{
+		public ButtonFocusTests()
+		{
+		}
+	}
+}
