+++ src/Core/tests/DeviceTests/Handlers/Editor/EditorFocusTests.cs
@@ -1,0 +1,12 @@
+﻿using Microsoft.Maui.DeviceTests.Stubs;
+
+namespace Microsoft.Maui.DeviceTests
+{
+	[Category(TestCategory.Editor)]
+	public class EditorFocusTests : FocusHandlerTests<EditorHandler, EditorStub, VerticalStackLayoutStub>
+	{
+		public EditorFocusTests()
+		{
+		}
+	}
+}
