+++ src/Core/tests/DeviceTests/Handlers/SearchBar/SearchBarFocusTests.cs
@@ -1,0 +1,12 @@
+﻿using Microsoft.Maui.DeviceTests.Stubs;
+
+namespace Microsoft.Maui.DeviceTests
+{
+	[Category(TestCategory.SearchBar)]
+	public class SearchBarFocusTests : FocusHandlerTests<SearchBarHandler, SearchBarStub, VerticalStackLayoutStub>
+	{
+		public SearchBarFocusTests()
+		{
+		}
+	}
+}
