+++ src/Core/tests/DeviceTests/Handlers/SearchBar/SearchBarHandlerTests.cs
@@ -297,25 +297,6 @@ namespace Microsoft.Maui.DeviceTests
 
 			await ValidatePropertyInitValue(searchBar, () => expected, GetNativeIsChatKeyboard, expected);
 		}
-
-		[Category(TestCategory.SearchBar)]
-		public class SearchBarTextInputTests : TextInputHandlerTests<SearchBarHandler, SearchBarStub>
-		{
-			protected override void SetNativeText(SearchBarHandler searchBarHandler, string text)
-			{
-				SearchBarHandlerTests.SetNativeText(searchBarHandler, text);
-			}
-
-			protected override int GetCursorStartPosition(SearchBarHandler searchBarHandler)
-			{
-				return SearchBarHandlerTests.GetCursorStartPosition(searchBarHandler);
-			}
-
-			protected override void UpdateCursorStartPosition(SearchBarHandler searchBarHandler, int position)
-			{
-				SearchBarHandlerTests.UpdateCursorStartPosition(searchBarHandler, position);
-			}
-		}
 #endif
 
 		[Category(TestCategory.SearchBar)]
