+++ src/Core/tests/DeviceTests/Handlers/SearchBar/SearchBarHandlerTests.iOS.cs
@@ -136,72 +136,6 @@ namespace Microsoft.Maui.DeviceTests
 			Assert.Equal(xplatCharacterSpacing, values.PlatformViewValue);
 		}
 
-		[Fact]
-		public async Task SearchBarFocusedIsWorking()
-		{
-			EnsureHandlerCreated(builder =>
-			{
-				builder.ConfigureMauiHandlers(handler =>
-				{
-					handler.AddHandler<VerticalStackLayoutStub, LayoutHandler>();
-					handler.AddHandler<SearchBarStub, SearchBarHandler>();
-				});
-			});
-
-			var layout = new VerticalStackLayoutStub();
-
-			var searchBar1 = new SearchBarStub
-			{
-				Text = "SearchBar 1"
-			};
-
-			var searchBar2 = new SearchBarStub
-			{
-				Text = "SearchBar 2"
-			};
-
-			layout.Add(searchBar1);
-
-			layout.Width = 100;
-			layout.Height = 150;
-
-			await InvokeOnMainThreadAsync(async () =>
-			{
-				var contentViewHandler = CreateHandler<LayoutHandler>(layout);
-				await contentViewHandler.PlatformView.AttachAndRun(async () =>
-				{
-					searchBar1.ToPlatform().Focus(new FocusRequest(false));
-					await searchBar1.WaitForFocused();
-					Assert.True(searchBar1.IsFocused);
-					Assert.False(searchBar2.IsFocused);
-				});
-			});
-		}
-
-		[Fact]
-		public async Task SearchBarUnFocusedIsWorking()
-		{
-			var searchBar = new SearchBarStub
-			{
-				Text = "SearchBar"
-			};
-
-			await InvokeOnMainThreadAsync(async () =>
-			{
-				var searchBarHandler = CreateHandler(searchBar);
-				await searchBarHandler.PlatformView.AttachAndRun(async () =>
-				{
-					searchBar.Handler.Invoke(nameof(IView.Focus), new FocusRequest(false));
-					await searchBar.WaitForFocused();
-					Assert.True(searchBar.IsFocused);
-
-					searchBar.Handler.Invoke(nameof(IView.Unfocus), new FocusRequest(false));
-					await searchBar.WaitForUnFocused();
-					Assert.False(searchBar.IsFocused);
-				});
-			});
-		}
-
 		double GetInputFieldHeight(SearchBarHandler searchBarHandler)
 		{
 			return GetNativeSearchBar(searchBarHandler).Bounds.Height;
@@ -213,16 +147,16 @@ namespace Microsoft.Maui.DeviceTests
 		string GetNativeText(SearchBarHandler searchBarHandler) =>
 			GetNativeSearchBar(searchBarHandler).Text;
 
-		static void SetNativeText(SearchBarHandler searchBarHandler, string text) =>
+		internal static void SetNativeText(SearchBarHandler searchBarHandler, string text) =>
 			GetNativeSearchBar(searchBarHandler).Text = text;
 
-		static int GetCursorStartPosition(SearchBarHandler searchBarHandler)
+		internal static int GetCursorStartPosition(SearchBarHandler searchBarHandler)
 		{
 			var control = searchBarHandler.QueryEditor;
 			return (int)control.GetOffsetFromPosition(control.BeginningOfDocument, control.SelectedTextRange.Start);
 		}
 
-		static void UpdateCursorStartPosition(SearchBarHandler searchBarHandler, int position)
+		internal static void UpdateCursorStartPosition(SearchBarHandler searchBarHandler, int position)
 		{
 			var control = searchBarHandler.QueryEditor;
 			var endPosition = control.GetPosition(control.BeginningOfDocument, position);
