--- src/Core/tests/DeviceTests/Stubs/StubPropertyChangedEventArgs.cs
@@ -1,17 +1,0 @@
-using System;
-
-namespace Microsoft.Maui.DeviceTests.Stubs
-{
-	public class StubPropertyChangedEventArgs<T> : EventArgs
-	{
-		public StubPropertyChangedEventArgs(T oldValue, T newValue)
-		{
-			OldValue = oldValue;
-			NewValue = newValue;
-		}
-
-		public T OldValue { get; }
-
-		public T NewValue { get; }
-	}
-}
