+++ src/TestUtils/src/DeviceTests/AssertionExtensions.cs
@@ -54,7 +54,6 @@ namespace Microsoft.Maui.DeviceTests
 			Assert.True(diff <= epsilon, $"Expected: {expected}. Actual: {actual}. Diff: {diff} Epsilon: {epsilon}.{message}");
 		}
 
-#if !TIZEN && PLATFORM
 		public static Task AssertHasContainer(this IView view, bool expectation)
 		{
 			// On Windows the `Parent` of an element only initializes when the view is added
@@ -145,12 +144,9 @@ namespace Microsoft.Maui.DeviceTests
 
 		public static bool IsExcludedWithChildren(this IView view) =>
 			view.ToPlatform().IsExcludedWithChildren();
-#endif
-
 
 		public static IDisposable OnUnloaded(this IElement element, Action action)
 		{
-#if PLATFORM
 			if (element.Handler is IPlatformViewHandler platformViewHandler &&
 				platformViewHandler.PlatformView is not null)
 			{
@@ -158,14 +154,10 @@ namespace Microsoft.Maui.DeviceTests
 			}
 
 			throw new InvalidOperationException("Handler is not set on element");
-#else
-			throw new NotImplementedException();
-#endif
 		}
 
 		public static IDisposable OnLoaded(this IElement element, Action action)
 		{
-#if PLATFORM
 			if (element.Handler is IPlatformViewHandler platformViewHandler &&
 				platformViewHandler.PlatformView is not null)
 			{
@@ -173,22 +165,14 @@ namespace Microsoft.Maui.DeviceTests
 			}
 
 			throw new InvalidOperationException("Handler is not set on element");
-#else
-			throw new NotImplementedException();
-#endif
 		}
 
 		public static bool IsLoadedOnPlatform(this IElement element)
 		{
-
-#if PLATFORM
 			if (element.Handler is not IPlatformViewHandler pvh)
 				return false;
 
 			return pvh.PlatformView?.IsLoaded() == true;
-#else
-			return true;
-#endif
 		}
 	}
 }
